Code completion (#36)
* CompletionItemProvider
* Set CompletionItemKind for Objects
* Fix Parser returns lowerCase Names
* add Return Value Type to Functions
* Fix Parser Problem with multiline commands.
